EXC-1207: Introduce E2E tests + ingest blocks in heartbeat (#21)
Heartbeats can now process responses, however the way these responses
are processed is brittle. The implementation of the heartbeat will be more robust with EXC-1208.

The focus of this PR was on adding an e2e test, which includes:

* A mock canister for providing fake blocks.
* A way to override the `blocks_source` in the bitcoin canister to retrieve fake blocks.
* A bash script containing a test
* CI integration

Needless to say, there are still lots of room for improvement that will happen in subsequent PRs,
including optimizations to CI to speed it up (EXC-1211) and refactoring our workflow to make it
easier to maintain (EXC-1212).