EXC-1209: Time-slicing within a single transaction. (#27)
A previous PR introduced time-slicing of a single block. However, with more experimentation
this wasn't sufficient, as there exist single transactions that are very large (thousands of
inputs and outputs), which can lead us to exceed the instructions limit.

This PR makes time-slicing more granular, where we now time-slice on each input/output of
a transaction, making the canister able to ingest very large transactions.