feat: start profiling the bitcoin canister (#43)
This MR kicks off the work for profiling the bitcoin canister. It adds a
script `profile.sh`
that runs the e2e test, and outputs some instruction counts into a file.

The idea is that this profiling will, in the few upcoming MRs, capture
the number of
instructions consumed by various operations (getting fee percentiles,
ingesting blocks, etc).

On every commit, the instruction counts file will be updated so that
it's clear to the reviewers
and in our history the impact of any change we make on the instructions
count. A CI test will
be added to enforce that the instruction coutns in
`e2e-tests/instructions_count.txt` are
always up-to-date.