Profile block ingestion
Track profiling stats related to block ingestion. For testing purposes,
the syncing_test had to be modified slightly by adding an extra block so
that it ingests more blocks.

The stats are already insightful. It's showing that, when ingesting a
block with thousands of inputs, almost the entire time is spent
computing tx ids. Making this more efficient will be one of the
optimizations that will be explored next.