refactor: introduce a new `Block` type. (#45)
The profiling results are evidence that computing txids is very
expensive.

Txids are computed in a number of places:

1. Replaying unstable blocks (in `get_balance` and `get_utxos`)
2. Computing fee percentiles
3. Ingesting blocks into the stable UTXO set

An obvious optimization is to cache txids to avoid recomputing them. To
be
able to do this effectively, this commit wraps the `bitcoin::Block` type
into a
new `Block` type. Having our own block type will give us the freedom to
cache
txids inside this struct in future commits.