feat: cache txids (#47)
Introduce a new `Transaction` type that wraps a `bitcoin::Transaction`,
and internally cache computations to `txid`.

We can see how the ingestion of block 1 became much faster. However,
block 1 is an extreme case because we
have a single transaction with a very large number of outputs, and this
isn't very common in real networks.
Nonetheless, it's a good optimization to have and it paves the way to
replace the implementation of `txid()`
with something that's more efficient.