optimize: make get_utxos more efficient. (#49)
In the presence of an unstable block that has thousands of outputs,
`get_utxos` can exceed the instruction limit. The new block that's added
to the syncing_test initially exceeded the instruction limit, and the
reason it exceeded the limit is because we kept deserializing all the
`added_utxos` for every input, which can amount to a lot of work if we
have a large number of them.

This commit changes the code so that we deserialize the `added_utxos`
only once at the very end.