feat: add new check in syncing test.
A new get_utxos call is added to the syncing test, which helps us both
test and profile an edge case where an address.has a large number of
stable UTXOs, but all of these UTXOs are removed in the unstable blocks.

This new call, as the profiling results show, is quite expensive. Future
commits will bring optimizations to get_utxos that will make this call,
among other get_utxos call, much cheaper.