optimize: make fee percentiles computation 40x faster (#56)
A bottleneck that is shared between the `get_utxos` and
`get_current_fee_percentiles` endpoints is fetching the transaction
output that is referenced by a particular outpoint.

In `get_current_fee_percentiles` this fetch is needed to get the fee of
a transaction, and in `get_utxos` it is needed to figure out the address
of the sender.

This commit introduces a `TxOutCache`, which provides fast access to all
the transaction outputs referenced by the unstable blocks. In this
commit, I've used that cache to optimize `get_current_fee_percentiles`,
and in a future commit I'll use that cache to optimize `get_utxos`.