optimize: make replaying blocks more efficient.
Replaying blocks is required in `get_utxos` and `get_balance`, where
we need to apply the unstable blocks on top of the stable UTXO set to
figure out the state at the specified number of confirmations.

This commit pre-computes a map of how each address in a block is affected
(what outpoints are added and removed). These can then be applied much
more quickly than doing that computation on the fly for every single
request.