feat: EXC-1231 properly compute `get_balance` while ingesting a block.
Prior to this commit, `get_balance` can return an incorrect result while
a block is being ingested, and that's because the UTXO set is in a state
where part of a block is written.

This commit makes ingesting blocks seem atomic. From the perspective of
`get_balance`, the balance reverts any changes made by the partially
ingested block, and only once the block is fully ingested is the balance
updated.

Notes:
* The amount of instructions required to ingest a block have regressed
  a bit, but that's expected given the additional complexity in tracking
  the partial state.

* We'll need to apply a similar logic as well to `get_utxo`, and that
  will be done in a follow-up commit.