refactor: make the UtxoSet state private. (#62)
In the replica implementation, the state was defined in one crate and
the implementation was defined in another. A disadvantage of that
approach is that all the state had to be public, and could potentially
be manipulated by outside modules.

Now that we're in a canister again, this MR makes most of the UtxoSet
state private, including the `Utxos` module and exposes a public
interface for consumers to use.