feat: introduce e2e test scenario 2.
Refactors the e2e tests into scenarios, and adds scenario 2 where we
test fetching the UTXOs of an address with a very large number of stable
UTXOs.

Scenario 2 currently fails as the call to `get_utxos` exceeds the
instruction limit. That will be fixed in a follow-up commit, and the
scenario will then be added into CI.

Note that there's a fair amount of boilerplate between the two scenarios.
I'll iteratively improve on that over time, and I'd rather not do that
now until the concept of scenarios, and what we require of them, is more
solidified.