fix: panic when removing an input with zero value. (#69)
There is an edge case where an address has UTXOs with positive values
and a UTXO with zero value. The positive UTXOs are all consumed,
bringing the address balance to zero, then the UTXO with zero value is
consumed.

This scenario cannot happen on mainnet, but can and has happened on
testnet, and without this fix would cause the canister to crash.