optimize: performance enhancement to `get_utxos`. (#71)
Optimizes `get_utxos` to enable scenario 2, which ran out of
instructions prior to this commit.

A big bottleneck in the previous iteration of this code is that we
fetched _all_ stable UTXOs after the provided offset. This means if an
address had 1M stable UTXOs, we'd fetch all of them and deserialize
them.

This implementation replaces that logic with iterators so that we only
deserialize the UTXOs we need. Additionally, the code has been
refactored to use a `Utxo` type, which is more clear to read.