feat: keep block in `unstable_blocks` until it is fully ingested.
Prior to this commit, the way stable blocks in `unstable_blocks` were
moved moved to the stable UTXO set was as follows:

1. Take a (newly stable) block from `unstable_blocks`.
2. Ingest the newly stable block into the `UtxoSet`.

Because of our time-slicing, steps 1 and 2 are not atomic, and the canister
can be in a state where the block is no longer part of the `unstable_blocks`,
and it hasn't been fully ingested yet.

This commit paves the way to make steps 1 and 2 atomic. It modified the
flow to the following.

1. Take a _reference_ to a block from the `unstable_blocks`.
2. Ingest the newly stable block into the `UtxoSet`.
3. Remove the block from `unstable_blocks`.

That way, the block remains in `unstable_blocks` until it is fully ingested
by the `UtxoSet`. In a follow-up commit, the `UtxoSet` will be modified

TBC