feat: keep block in `unstable_blocks` until it is fully ingested. (#72)
Prior to this commit, the way stable blocks in `unstable_blocks` were
moved to the stable `UtxoSet` was as follows:

1. Take a (newly stable) block from `unstable_blocks`.
2. Ingest the block into the `UtxoSet`.

Because of our time-slicing, steps 1 and 2 are not atomic, and the
canister can be in a state where the block is no longer part of the
`unstable_blocks`, and it hasn't been fully ingested yet.

This commit paves the way to make steps 1 and 2 atomic. It modifies the
flow to the following.

1. Take a _reference_ to a block from the `unstable_blocks`.
2. Ingest the block into the `UtxoSet`.
3. Remove the block from `unstable_blocks`.

This flow as it currently stands is still not atomic, but to pave the
way for atomicity, I'd like the block to remain in `unstable_blocks`
until it is fully ingested by the `UtxoSet`. In a follow-up commit,
changes to the `UtxoSet` will guarantee the atomicity of these three
operations.