feat: (EXC-1231) make block ingestion atomic.
While a block is being ingested, the `UtxoSet` now reverts all the partial changes that are made by that block until it is fully ingested. In other words, the `UtxoSet` only exposes the changes by the ingested block after it's entirely ingested, and never at points where it's partially ingested.