feat: add endpoint to know if a block is being ingested.
This change is temporary to allow us to upgrade the canister with the atomicity changes safely.

To include the atomicity changes safely, we need to upgrade the canister when it is _not_ in the middle of ingesting a block.

This commit introduces an endpoint for us to know if that is the case so that we can upgrade correctly.

Once the upgrade is done, this change will be reverted.