fix: improve the fee percentile computation (#88)
The current implementation handles the case when there are fewer fees
than "buckets" well. However, it does not provide accurate results in
the other case. Since the Bitcoin canister looks at the percentiles
(i.e., 100 buckets) of 10,000 transactions at a time, the later case is
more important.

Concretely, for the input \[1, 2, ..., 10_000\], the current
implementation returns \[1, 101, 201, ..., 9901\]. There are several
issues with this result:

* The returned median is 4901 instead of 5000.
* The smallest value (1) is returned, which is not the first percentile.
* The 100th percentile does not correspond to the largest value
(10_000).

This PR fixes these issues.