feat: [EXC-1283] script to compute the address balances.
Rather than computing the entire state from scratch by feeding all the
blocks to the canister, we'll now rely on Bitcoin's chainstate database.
This will allow us to compute the whole state in a matter of hours, as
opposed to weeks or months.

This commit introduces one of several scripts that will be used to read
data from the chainstate database and compute various parts of the
canister's state.