feat: [EXC-1286] Validate the bytes passed to the uploader canister.
To ensure the correctness of the bytes uploaded via the `uploader`
canister, the `uploader` canister is now initialized with a list of
the sha2 hashes of each chunk that'll be uploaded. The `uploader` then
verifies the hash of each chunk prior to writing it into stable memory.

A script for compute the chunk hashes is also included.