feat: Use docker for reproducible canister builds (#99)
Add a Dockerfile that allows for reproducible canister builds for both
the bitcoin as well as the uploader canister. To share most of the
parts, the build script for the bitcoin canister is moved to
`scripts/build-canister.sh` that accepts as an argument the name of the
canister that should be built.