fix: load all supported addresses from the utxo dump file (#103)
When pre-computing the state we rely on a 3rd party tool that outputs a
text dump of the UTXOs. Not all the addresses we support are emitted by
this tool, so as a result we don't include all the addresses we support
in our state.

This commit fixes this issue by adding a fallback mechanism. If parsing
the address fails (e.g. it doesn't exist), we attempt to parse the
script of the UTXO directly.