feat: script to compute main state struct
Adds a script to compute the main `State` struct of the canister. This
script requires loading the canister from a file, and to enable that
this commit adds a feature to the canister, `file_memory`, that allows
initializing the canister from a file.

Because this script needs to the `file_memory` feature and none of the
other scripts, it had to be in own crate and in its own workspace.