feat: script to compute main state struct (#105)
Adds a script to compute the main `State` struct of the canister. This
script requires loading the canister from a file, and to enable that
this commit adds a feature to the canister, `file_memory`, that allows
initializing the canister from a file.

Because this script needs to use the `file_memory` feature and all of
the other scripts in the `state-builder` crate require _not_ using the
`file_memory` feature, it had to be in own crate and in its own
workspace.