feat: Add a flag to disable bitcoin apis (#120)
This PR adds a flag that allows us to enable/disable access to the
bitcoin apis provided by the canister. The flag is set to have a default
value of `true` to not disrupt the current operations of the canister.
It's also added in the canister's `Config` so that it can eventually be
set through a proposal that triggers the `set_config` endpoint of the
canister. For now, if we need to disable the apis, we can set the flag
to `Disabled` in the canister's post upgrade hook.