fix: fix bug in retrieving the caller in the set_config endpoint.
Problem:
When calling the set_config API, the following error is returned:

```
"ic0_msg_caller_size" cannot be executed in reply callback mode
```

The reason here is that the caller is fetched after making an async
request, which is not something that we allow.

Solution:
Store the caller in a variable before making the async request.