feat: [EXC-1374] add script for running benchmarks.
In a previous commit, a benchmarks canister was introduced to help
benchmark specific parts of the code.

This commit includes a script to run these benchmarks. The script relies
on `drun`, which is configured to run benchmarks with practically no
instruction limit. For now, it is assumed to be available in the caller's
path.

Running `./benchmarks/run.sh`, returns the following:

```
(113_061_801_763 : nat64)
```

In other words, the one benchmark we currently have takes 113B
instructions to run (!!)