optimize: cache block hash computations to speed up block insertions
Our `insert_300_blocks` benchmarks revealed that inserting the first
300 testnet blocks takes ~113B cycles. Looking more closely at this, it
turns out that the vast majority of these cycles were spent computing
block hashes over and over again.

Caching the block hashes reduced the instruction count to the following:

```
(6_366_348_305 : nat64)
```

As in, the benchmark is now ~95% faster. Note that this doesn't really
 translate to mainnet, because the logic that was computing the block
hashes aggressively was specific to the testnet/regtest networks, but
now we have good tools to triage other performance issues if need be.

What this actually fixes is testnet syncing. Currently, if we were to
deploy a bitcoin testnet canister and have it sync from scratch, it'll
run out of cycles due to this computation (even on a system subnet), and
if we ever disable syncing on the testnet canister and re-enable it,
there would be a similar risk, so this optimization is still impactful.