chore: upgrade stable structures to version 0.5.2 (#176)
The API for stable structures changes quite significantly between the
two versions, so the change is relatively involved.

Due to a (now known) performance regression in stable structures >=
0.4.0, `address_utxos` have been migrated to use a `Blob` in the
`BTreeMap` rather than an `AddressUtxo`. Once the performance regression
is fixed, we'll use `AddressUtxo` as the type again.

There's a minor performance regression, mostly due to the conversions
between `Blob` and `Vec`. These will be addressed once the regression
mentioned above is addressed, but given that it's minor, it's not a
blocker for the upgrade.