fix: fix mocking concurrent http requests with transform functions in ic-http (#180)
This PR fixes storing transform functions in a thread-local storage
which was preventing to test concurrent http requests.

Previously it was incorrectly calculating the name of the function,
since it was called inside `create_transform_context` function with a
specific type of `TransformFn`, which loses the name of the function (in
contrast to a template parameter) and only stores the function signature
(which is always the same).

To fix this it had to convert `create_transform_context` to a template
and store a boxed function pointer in a thread-local storage. Added
tests to make sure function names are stored properly now.