feat: create a watchdog canister to fetch Bitcoin block data from explorers (#177)
This PR introduces a watchdog canister for tracking bitcoin canister
health, which is capable of fetching bitcoin block data from different
remote APIs.

The implementation uses the `BitcoinBlockApi` enum variant to describe
each remote API provider, along with the corresponding `fetch_data()`
function that may require sending one or several HTTP requests. The
`Endpoint` enum variant is used to describe each HTTP request, which
allows for storing all configuration in one place, including the URL,
the transform function exposed by the canister, and its respective
implementation. This approach allows for all of the canister's exposed
methods to be located at the top level in lib.rs, while the actual
implementation of the transform function is stored next to the URL and
tests to improve readability.

The implementation uses `ic-http` crate, which allowed for testing API
responses on mocked data.

Since this PR is already quite large, it is separated from processing
and serving the results to the bitcoin_canister, which will follow in
separate PRs.