perf: calculate the main chain height more efficiently (#237)
The Bitcoin testnet's metrics endpoint ran out of cycles because it had
a large number of unstable blocks, and computing the main chain height
under those conditions was very expensive.

This commit makes that computation more efficient. Prior to this change,
the `get_metrics` benchmark took `163_059_419` cycles. Now it takes
`8_102_095`.