perf: skip next block headers if they are already inserted.
As the Bitcoin canister is syncing, it can encounter the following
pattern if it's falling far behind:

```
unstable_state = {
  blocks: {A, B, C};
  next_block_headers: {D, E, F}
}

// Calls get_successors, and gets the following response:
get_successors_response {
  blocks: {D},
  next_block_headers: {E, F}
}
```

In this scenario, the canister processes block headers `E` and `F` twice,
which is unnecessary computation.

This commit adds logic to check if a block header has already been added
to the unstable blocks and, if yes, skips it. A benchmark has been added
to prevent this from regressing.

Before the optimization:

```
insert_block_headers_multiple_times
                        time:   [118.21 B Instructions 118.21 B Instructions 118.21 B Instructions]
```

After the optimization:

```
insert_block_headers_multiple_times
                        time:   [12.055 B Instructions 12.055 B Instructions 12.055 B Instructions]
                        change: [-89.802% -89.802% -89.802%] (p = 0.00 < 0.05)
                        Performance has improved.
```