fix: add a bound on the length of the unstable chain in testnet/regtest
In production, we set a `stability_threshold` of ~144. However, when
we consider the stability of the chain, we consider it weighted by the
blocks' difficulties. This isn't a problem in the normal case, where
block difficulties are very similar.

There are scenarios specific to the Bitcoin testnet, however, that can
make this logic problematic. The difficulty in the Bitcoin testnet can
be reset to the minimum in case a block hasn't been found for 20 minutes.

Consider the following scenario:

* Assume a `stability_threshold` of 144.
* The anchor at height `h` has difficulty of 4642.
* The anchor will be marked as stable if the difficulty-based depth of
  the successor blocks is `stability_threshold * difficulty(anchor)` = 668,448
* The difficulty is reset to the minimum of 1.
* The canister will now need to maintain a chain of length 668,448 just to
  mark the anchor block as stable!

Very long chains can cause the shadow stacks to overflow, resulting in a
broken canister.

The pragmatic solution in this case is to bound the length of the chain. If
there's only one chain and it starts exceeding a certain length, we assume
that the anchor is stable even if the difficulty requirement hasn't been
met.

This scenario is only relevant for testnets, so this addition is safe and
has not impact on the behavior of the mainnet canister.