fix: drop next block headers above a certain instructions threshold
Inserting next block headers can be expensive and cause us to go over
the 50B instruction limit that's available for the heartbeat. Not only
does this cause the canister to stop syncing, but it also reduces the
finalization rate of the subnet significantly as the subnet repeatedly
tries to run a heartbeat that executes 50B instructions (which takes
about 10 seconds).

A simple solution is, ff we are approaching the instruction limit while
processing the next block headers, we ignore the remaining. They'll be
delivered by the adapter in the following call to `get_successors`.