fix: failing scenario 1 test (#260)
The e2e scenario 1 test started failing lately. It appears that the
behavior of dfx has changed and the error messages being returned, which
we use in the test, have changed. I don't quite understand how the
behavior of one particular version of dfx has changed, unless that
version has been modified. I'm following up with the SDK team to
understand that. In any case, this fixes the failures we're seeing.