fix: deserialize `BlockTree` iteratively (#258)
# Problem
We're now running into a situation where the Bitcoin testnet has ~125k
unstable blocks due to a difficulty reset.
Recursively deserializing these blocks, which we do in `post_upgrade`,
causes a stack overflow.

# Solution
Rewrite the `BlockTree` deserialization to be iterative rather than
recursive. Running a test with the current production state confirms
that it fixes that stack overflow issue.