docs: Update development-guide.md
Github doesn't have a good flow for handling releases for two different artifacts. For example, we cannot have a release marked as the latest release for the Bitcoin canister and a different release marked as the latest release for the watchdog canister.

Instead, we now always release the Bitcoin canister and watchdog canister together, and they'll all need to be re-deployed in the case of a new release.

If there happens to be only modifications in one of the binaries, say, the watchdog canister, then in theory the Wasm for the Bitcoin canister would be unchanged, and would therefore not need a release.