chore: reduce min_explorers to 2 in tests setup (#270)
This PR fixes `watchdog_health_status` test by dropping `min_explorers`
requirement to 2 in test config.

The reason for the failure is that there's no IPv4 support in dfx for
system subnets, which causes only 2 explorers (supposedly with IPv6) to
have http data which was not enough for the test to pass (old
requirement was to have at least 3 explorers). Dropping the requirement
temporarily fixes the test.

This PR is a short-term workaround. Long-term it'd require either a
support for IPv4 or more explorers with IPv6.

Response example where `api_blockcypher_com_mainnet` and
`chain_api_btc_com_mainnet` have responses:
```
  record {
    height_target = null;
    height_status = variant { not_enough_data };
    explorers = vec {
      record {
        height = null;
        provider = variant { api_blockchair_com_mainnet };
      };
      record {
        height = opt (818_978 : nat64);
        provider = variant { api_blockcypher_com_mainnet };
      };
      record { height = null; provider = variant { blockchain_info_mainnet } };
      record { height = null; provider = variant { blockstream_info_mainnet } };
      record {
        height = opt (818_978 : nat64);
        provider = variant { chain_api_btc_com_mainnet };
      };
    };
    height_source = opt (818_978 : nat64);
    height_diff = null;
  },
```

Links:
- PASS
https://ready.chair6.net/?url=https%3A%2F%2Fapi.blockcypher.com%2Fv1%2Fbtc%2Fmain
- PASS
https://ready.chair6.net/?url=https%3A%2F%2Fchain.api.btc.com%2Fv3%2Fblock%2Flatest
- FAIL
https://ready.chair6.net/?url=https%3A%2F%2Fapi.blockchair.com%2Fbitcoin%2Fstats