fix: make the bitcoin testnet watchdog less sensitive (#277)
Sets the threshold for blocks ahead/behind from +/- 2 to +/- 100 for
bitcoin testnet given that bitcoin testnet often has surges in blocks
when the difficulty is low.

The configs of the watchdog canister were not being maintained in this
repo, and instead were only maintained in the NNS proposals. After a
discussion with @maksymar, we agreed to maintain the configurations here
instead to make them easier to review and track.

As a result, the watchdog canister has now been updated to require a
bitcoin network when being initialized, and based on that network it'll
use the hard-coded bitcoin mainnet/testnet config.