fix (EXC-1590): make burning cycles configurable (#286)
# Problem
The latest version of the Bitcoin canister automatically burns all
cycles it receives. DFX is unable to upgrade to that version since
deploying it on a local app subnet immediately results in it being
uninstalled.

# Solution
Make burning cycles configurable, such that DFX can disable it when
deploying the Bitcoin canister locally.