chore: update threshold of watchdog testnet canister (#295)
With recent block surges it's becoming more common for the bitcoin
testnet to be ahead of external block explorers. To avoid disabling the
bitcoin testnet API for extended periods of time, this commit makes the
watchdog less sensitive to differences in height between the bitcoin
testnet canister and external explorers, making it less likely for it to
disable the Bitcoin API for developers.