feat: add config to eagerly evaluate fee percentiles (#303)
Fee percentiles are typically evaluated lazily, meaning that they are
not computed unless they're requested. This can be problematic in local
development where there are very few transactions. Consider the
following scenario in local development:

1. User mines a block with a non-coinbase transaction
2. User mines a block with only a coinbase transaction

If developers now call the `get_current_fee_percentiles` endpoint,
they'll get an empty vector. This is incorrect, as there has already
been one transaction with fees that can be used to compute the fee
percentiles. This behavior happens because locally all blocks are marked
as stable immediately, and we only evaluate fee percentiles in unstable
blocks.

To address this issue, this commit adds a new configuration
`lazily_evaluate_fee_percentiles`, which allows us to configure the
canister whether fee percentiles are lazily evaluated or not. For local
development environment, they'd be evaluated eagerly, and lazily
otherwise.

See [this forum
post](https://forum.dfinity.org/t/get-current-fee-percentiles-returns-empty-array-after-mining-a-block-with-no-transactions)
for an example of a user running into this error.