optimize: reduce memory footprint of pre_upgrade by 50% (#306)
Another optimization where we, instead of serializing the entire `Block`
struct, we only serialize the `block` field inside it, which actually
has all the information we need to reconstruct the block when
deserializing.

For pragmatic reasons, I disabled the test that ensures that a canister
can self-upgrade and instead we only have the test that ensures we can
upgrade the latest release. Supporting both in the same code proved to
be more complicated than expected.

Once the canister is released, the `post_upgrade` will be updated the
canister self-upgrade test will be re-enabled.