chore: more enhancements for bootstrapping (#316)
1. `post_upgrade` hook now take an optional configuration. This allows
the upgrade proposal after bootstrapping to set appropriate configs,
particularly fees.
2. The `post_upgrade` needs to work with the bootstrapped canister and
the current bitcoin mainnet canister. The bitcoin mainnet canister still
has an old `pre_upgrade` hook. To address this, a `legacy_preupgrade`
feature has been added. When upgrading the bootstrapped testnet
canister, this feature won't be enabled, but when upgrading the bitcoin
mainnet canister, this feature will need to be enabled. This has the
unfortunate consequence that in the upcoming release the bitcoin testnet
and mainnet canister wasms won't match.