fix: do not include canbench in production (#317)
## Problem
Running `pre_upgrade` in the bootstrapping scripts fails with an error
that `stable64_size` is only available in canisters.

## Solution
The bootstrapping scripts do not run the canister code as Wasm, and do
not have access to IC-level APIs. Upon further investigation, the reason
`stable64_size` was needed was because we were accidentally including
`canbench` into the compilation of the canister code - it should only be
included when we're benchmarking.

To fix this, the `build-scripts.sh` has been updated to build using
`--package` rather than `--bin` - that actually builds the canister from
the correct root `Cargo.toml` file and therefore doesn't include any
additional features that were required by other packages.

I used this opportunity to add a test using `PocketIC`.