fix: uploader canister compilation (#323)
The upload canister compilation was failing with the following error:

```
error: Only features sync,macros,io-util,rt,time are supported on wasm.
   --> /Users/ielashi/.cargo/registry/src/index.crates.io-6f17d22bba15001f/tokio-1.39.3/src/lib.rs:467:1
    |
467 | compile_error!("Only features sync,macros,io-util,rt,time are supported on wasm.");
    | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
```

This is due to a dependency on `tokio`, which for some reason `sha256`
crate needs. This commit replaces `sha256` with `sha2` which has no such
dependencies and fixes the compilation.

It turns out there are not tests to ensure the compilation of the upload
canister is correct. These should be added, but better done in a
separate PR.