feat: Add a flag to allow for uploading chunks from the end of the list (#335)
This adds a flag to the uploader binary to allow for choosing whether
the upload starts from the first chunk (default) or the last chunk. This
can be handy to speed up the upload of state since once can start two
instance in parallel, one uploading from the first chunk and the other
from the last one.