fix(validation): add timestamp validation check testnet4 (#402)
[XC-387](https://dfinity.atlassian.net/browse/XC-387?atlOrigin=eyJpIjoiMGViYWMzYjk5MmY1NGIyOThlZDdiMGE0YWZmMzIyNjUiLCJwIjoiaiJ9):
Previously, validation of testnet4 block headers was skipped because the
time returned by `ic_cdk::time()` within benchmarks was too far in the
past, causing the condition that header timestamp must be less than 2
hours in the future to fail. PR
https://github.com/dfinity/bitcoin-canister/pull/401 addressed this by
introducing support for mocking time during benchmarks. With this
change, timestamp validation for testnet4 headers can now be re-enabled.

[XC-387]:
https://dfinity.atlassian.net/browse/XC-387?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ