refactor(types): remove useless BitcoinAddress::from_str safety check (#408)
Due to a bug in early versions of the bitcoin crate (v0.28), it was
possible to convert an invalid script into an address using
`Address::from_script`. To get around this bug, we used to convert the
address to a string, and verify that this string is a valid address
before returning it. This bug was raised in
https://github.com/rust-bitcoin/rust-bitcoin/issues/995 and has been
fixed since then
(https://github.com/rust-bitcoin/rust-bitcoin/pull/1039). This PR
removes the unnecessary conversion to string and outdated references to
the issue.