refactor: replace RefCell<Option<T>> with OnceCell<T> (#431)
`OnceCell` is better suited to implement "initializing once" behavior
required for caching `BlockHash` and `Txid`.

Another important difference is that it is possible to return a direct
reference to the inner value with `OnceCell`, but not with `RefCell`
because of its dynamic borrow. This turns out to be handy at
implementing the cache for unstable blocks.