fix(watchdog): add cycles for http requests when in an application subnet (#439)
The watchdog canister was assumed to be deployed in a system subnet,
therefore no cycles were attached to make http requests. The watchdog
canister for the Dogecoin staging canister is actually deployed in an
application subnet. This PR attaches cycles to http request in case the
watchdog canister is deployed in an application subnet.