fix: Remove custom PartialOrd implementation for Utxo type (#451)
The previous custom implementation treats Utxos as Outpoint for
comparison purpose, but it still derives standard PartialEq and Eq
instances, which would cause a subtle bug for two Utxo values that have
equal outpoint but different height or value. In real world cases this
never happens but still, having this inconsistency presents a risk.

This custom implementation is not very helpful either because:
1. The cost it saves in comparison operation is marginal.
2. It silently hides problem when there are two Utxos with equal
outpoint but different height/value.

Therefore it is best to remove this custom implementation and use
standard derive instead.