feat(watchdog)!: add `health_status_v2` endpoint (#450)
[DEFI-2494](https://dfinity.atlassian.net/browse/DEFI-2494?atlOrigin=eyJpIjoiYWY2YTM0NTM2YmZlNDdjYTlkODIzYjRlNzNhMmVlMGEiLCJwIjoiaiJ9)

The following commit introduced an unexpected breaking change on the
`health_status` endpoint:
https://github.com/dfinity/bitcoin-canister/commit/2e9d8929996592182003a03244e1adf065273e90
. This is because new Dogecoin providers were added to the
`bitcoin_block_api` enum.

This PR reverts the addition of Dogecoin providers in the
`bitcoin_block_api` enum and introduces a new `health_status_v2`
endpoint to the canister which enables adding new explorers in the
future without introducing breaking changes to the `health_status_v2`
endpoint. The `health_status` endpoint is now marked as DEPRECATED.

This PR ensures that no breaking changes are introduced to the
`health_status` endpoint compared to the version of the [Watchdog
canister](https://dashboard.internetcomputer.org/canister/gatoo-6iaaa-aaaan-aaacq-cai)
that is currently running in production.

However, this PR introduces a breaking change in the `get_config`
endpoint by changing the type of the `explorer` field from
`bitcoin_block_api` to `text`, which is acceptable since it is a debug
endpoint.

Additionally, this PR introduces a new metric to track the number of
times the deprecated `health_status` endpoint is called by other
canisters.

[DEFI-2494]:
https://dfinity.atlassian.net/browse/DEFI-2494?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: gregorydemay <112856886+gregorydemay@users.noreply.github.com>