chore(watchdog): remove deprecated method counter (#455)
The counter `DEPRECATED_METHOD_CALLS` was added to count the number of
calls made to the deprecated `health_status` endpoint. However, this
endpoint is actually a query endpoint which prevents us from modifying
the state and being able to increment the counter when the endpoint is
called. This PR removes the code related to this counter.